requirements.txt
affine==2.4.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
cfgv==3.5.0
click==8.3.3
cligj==0.7.2
distlib==0.4.0
filelock==3.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.19
idna==3.13
iniconfig==2.3.0
librt==0.9.0
mypy==1.20.2
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pathspec==1.1.1
platformdirs==4.9.6
pluggy==1.6.0
pre_commit==4.6.0
pyarrow==24.0.0
Pygments==2.20.0
pyparsing==3.3.2
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6_sip==13.11.1
pytest==9.0.3
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-discovery==1.2.2
python-dotenv==1.2.2
PyYAML==6.0.3
rasterio==1.5.0
ruff==0.15.12
scipy==1.17.1
six==1.17.0
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
virtualenv==21.3.0
